packages/vision/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.26.2
torch==2.1.1
torchvision==0.16.1
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
